fastapi/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
python-dotenv==1.0.1
pydantic==2.8.2
motor>=3.0.0
starlette==0.37.2
requests==2.32.3
httpx==0.27.0
PyYAML==6.0
itsdangerous==2.1.2
uvicorn
jinja2
